Fix object_info field alignment — dimensions are 32-bit, not 16-bit
Per Carnage3D StyleData.cpp:720-722 each G24 object_info record is:
  SI32 width, SI32 height, SI32 depth,
  I16 baseSprite, I16 weight, I16 aux,
  I8 status, I8 numInto,
  UI16 into[numInto]

Our parser (mirroring the web project's TS parser, which doesn't
actually use BaseSprite because it hard-codes obj→sprite mapping)
was reading dimensions as 16-bit. That shifted every subsequent
field by 6 bytes — BaseSprite was reading the high half of Height,
Weight was reading low half of Depth, etc. So every object's
"BaseSprite" came out as some unrelated bit pattern, and every
placed object rendered with the same wrong sprite (the flame-like
shapes covering grass tiles in the user's screenshot).

Fix: G24ObjectInfo dimensions now `int`; G24Reader.ReadObjectInfo
reads them as ReadInt32LittleEndian, advancing 12 bytes for dims +
the 8 bytes of remaining fields = 20 bytes per record (was 14).

Tests still pass (12/12).

+++ src/GTA1MapEditor.Core/Models/G24Models.cs
@@ -58,9 +58,10 @@ public sealed record G24SpriteDelta(ushort Size, uint Offset);
 
 public sealed class G24ObjectInfo
 {
-    public short Width { get; set; }
-    public short Height { get; set; }
-    public short Depth { get; set; }
+    // Per Carnage3D StyleData.cpp:720-722 dimensions are signed 32-bit, not 16-bit.
+    public int Width { get; set; }
+    public int Height { get; set; }
+    public int Depth { get; set; }
     public ushort BaseSprite { get; set; }
     public ushort Weight { get; set; }
     public ushort Aux { get; set; }
